2010-01-28  Steve Block  <steveblock@google.com>
        Reviewed by David Levin.

        Use long long rather than int64 in V8Binding.h
        https://bugs.webkit.org/show_bug.cgi?id=34270

        No new tests, build fix only.

        * bindings/v8/V8Binding.h: Modified.
        (WebCore::toInt64): Returns long long

Canonical link: https://commits.webkit.org/45405@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54038 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/bindings/v8/V8Binding.h
@@ -112,9 +112,9 @@ namespace WebCore {
         return static_cast<float>(value->NumberValue());
     }
 
-    inline int64 toInt64(v8::Local<v8::Value> value)
+    inline long long toInt64(v8::Local<v8::Value> value)
     {
-        return static_cast<int64>(value->IntegerValue());
+        return static_cast<long long>(value->IntegerValue());
     }
 
     // FIXME: Drop this in favor of the type specific v8ValueToWebCoreString when we rework the code generation.
